fix warnings on crontab files

+++ require/softwares/AllSoftware.php
@@ -96,7 +96,7 @@ class AllSoftware
             'EXCLUDE_ARCHIVE_COMPUTER' => 'EXCLUDE_ARCHIVE_COMPUTER'
         ];
 
-        $configValues = look_config_default_values($configToLookOut)['ivalue']['EXCLUDE_ARCHIVE_COMPUTER'];
+        $configValues = look_config_default_values($configToLookOut)['ivalue']['EXCLUDE_ARCHIVE_COMPUTER'] ?? '';
 
         $sql = "SELECT CONCAT(n.NAME,';',p.PUBLISHER,';',v.VERSION) as identifier, 
                 s.VERSION_ID, s.NAME_ID, s.PUBLISHER_ID, 
@@ -137,7 +137,7 @@ class AllSoftware
             $i++;
         }
 
-        if ($unlinked_hids >= 1) {
+        if (isset($unlinked_hids) && $unlinked_hids >= 1) {
             $sql_del = "DELETE FROM software WHERE HARDWARE_ID IN (%s)";
             $arg_del = implode(",", $unlinked_hids);
             $result = mysql2_query_secure($sql_del, $_SESSION['OCS']["writeServer"], $arg_del);
@@ -158,7 +158,7 @@ class AllSoftware
                 $i++;
             }
 
-            if ($unlinked_ids >= 1) {
+            if (isset($unlinked_hids) && $unlinked_ids >= 1) {
                 $sql_del = "DELETE FROM $table WHERE ID IN (%s)";
                 $arg_del = implode(",", $unlinked_ids);
                 $result = mysql2_query_secure($sql_del, $_SESSION['OCS']["writeServer"], $arg_del);
@@ -179,7 +179,7 @@ class AllSoftware
                 $i++;
             }
 
-            if ($unlinked_ids >= 1) {
+            if (isset($unlinked_hids) && $unlinked_ids >= 1) {
                 $sql_del = "DELETE FROM `software_categories_link` WHERE ID IN (%s)";
                 $arg_del = implode(",", $unlinked_ids);
                 $result = mysql2_query_secure($sql_del, $_SESSION['OCS']["writeServer"], $arg_del);
